sdcc compiles everything! Need to build new object files for the given ones.

+++ lab7/delay.c
@@ -1,14 +1,14 @@
 #include "delay.h"
 
-void delay_ms(const double t) {
-    const int cycles = (t * (1. / 1000.) * (double)CPUCLK) + 1;
+void delay_ms(const float t) {
+    const int cycles = (t * (1. / 1000.) * (float)CPUCLK) + 1;
 
     int i;
     for (i = 0; i <= cycles; ++i) {} // wait
 }
 
-void delay_us(const double t) {
-    const int cycles = (t * (1. / 1000000.) * (double)CPUCLK) + 1;
+void delay_us(const float t) {
+    const int cycles = (t * (1. / 1000000.) * (float)CPUCLK) + 1;
 
     int i;
     for (i = 0; i <= cycles; ++i) {} // wait
