+++ lab7/delay.h
@@ -3,7 +3,7 @@
 
 #define CPUCLK  12250000    // SYSCLK in Hz
 
-void delay_ms(const double t);
-void delay_us(const double t);
+void delay_ms(const float t);
+void delay_us(const float t);
 
 #endif
